// server/vance-addon-brain-wiki/client/src/generated/wiki/WikiView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.wiki.WikiView
*/

import { WikiSpaceView } from './WikiSpaceView';
import { WikiPageView } from './WikiPageView';

export interface WikiView {
  folder: string;
  title?: string;
  description?: string;
  mainPagePath?: string;
  mainPageId?: string;
  indexPagePath?: string;
  indexPageId?: string;
  spaces: WikiSpaceView[];
  pages: WikiPageView[];
}