// server/vance-addon-brain-wiki/client/src/generated/wiki/WikiRebuildResponse.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.wiki.WikiRebuildResponse
*/

export interface WikiRebuildResponse {
  folder: string;
  rootIndexPath: string;
  backlinksPath?: string;
  indexCount: number;
  pageCount: number;
  spaceCount: number;
}