// server/vance-addon-brain-wiki/client/src/generated/wiki/WikiResolveResponse.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.wiki.WikiResolveResponse
*/

export interface WikiResolveResponse {
  target: string;
  exists: boolean;
  ambiguous: boolean;
  slug: string;
  path?: string;
  id?: string;
  space?: string;
  createSpace: string;
  createPath: string;
}